Fixed bug 131665: Wrong cause of death when cursed wands explode.
Note that the message given if you are killed directly by the explosion
from a cursed wand is still far from ideal. This is because explode()
detects whether to say the normal "caught himself in his own ..." (or her)
depending on the sign of <type>. We can't use -type in wand_explode when
hero_broke is false as we would like to because ZT_MAGIC_MISSILE is zero.

For now, I have just added a FIXME. A proper fix will take much more
thought.