GTK: Improved move and fight key handling.
Version 0.0.6E4F2 has a slight anomaly in that the keypad arrow keys
etc. can be mis-interpreted as letters which select menu items. This
has become a major anomaly in the current CVS version since I extended
the arrow key handling to include the non-keypad keys.

This fixed that problem by only doing arrow key handling for the default
key press handler and also correctly deals with the effects of num_pad.

The fight command has been re-activated and tested on both Linux and
Windows 95 with no problems apparent.