Subverted setguid check in GTK+ version 1.2.9 and above.
This is tested under Linux against GTK+ version 1.2.9 and
the underlying code to hide priveleges has been tested on
Solaris 2.5, Solaris 8, FreeBSD 4.2-STABLE and Compaq Tru64 v5.1
(I don't have access to GTK+ on these boxes to do any more).

The code assumes that all versions of UNIX that it can hide
privileges for (and there are many that it cannot) have
POSIX types (uid_t and gid_t). If this proves not to be the
case, this could be solved.

It would have been possible to keep all the hacking in one
file, but it would have made us technically break the Linux
documented API for system calls, so I decided it was better
to create a new file, unixres.c