Fixed bug 421136: engraving with wands of draining.
For the development branch only, I have also changed the messages
generated when engraving with a wand of draining. Depending on whether
the existing message is wiped completely or just degraded the messages
will match the ones from wands of cancellation/make invisible or wands
of polymorph respectively. This will make identifying wands of draining
slightly more difficult for most players, and significantly more difficult
for illiterate players.