Fixed bug 422153: inaccessible spells after first 52.
Spell menus no longer specify the accelerator (leaving the window system
to do this). This frees us from the contraints of 52 accelerators. This
has also meant a very slight change in that the "swap <> with ..." mesg
now uses the spell name since the accelerator is not known.

In traditional menustyle, the spell key is used to allow spells to be
cast without needing to display a menu. For this purpose, the keys are
extended to a-z A-Z 0-9 (62 in all). These keys are displayed as a
new column in the spell menu in this menu also for clairity.

Note: If we ever exceed 62 spells, the traditional menustyle will still
work but the 63rd, and beyond, spell will need to be accessed via the
menu (using the '?' command).