GTK: Added session save file.
Currently the only part of the session that is saved is the postion and
size of the standard windows: main, radar and permemant inventory.

I am planning on adding a new dialog to control user preferences which
are not part of the game options including the facility to specify
connections to servers for GtkHack. Such preferences will be saved as
part of the session data and so avoid needing to extend the game to
add many new options (it would not be possible for GtkHack to store
connection data via the game core anyway).